Moved and update the import list file
In the dependencies folder for the CompStart Python CLI tool was a file named import_list.txt. It was created to store all the import statements needed for the main Compstart script and all the dependency scripts.

Since it is no longer really needed, this file was moved to /devenv/data/old_data. It was also updated to reflect the naming change of the scripts from commit 3c14bb9 to keep everything consistent.