Changed CompStart.py to be in production
As noted in issue #43:

Tested copying the CompStart folder created by PyInstaller to my ~/Documents folder and received the same error. Looked more carefully at the error and realized the issue is a logic or code one.

The error is that in CompStart.py, there's the global variable is_prod which is used to determine the path to the config folder. It was set to False, but changed it to True.

May need to figure out a better way to handle changing locations of the config folder.