Updated the path to config for testing environment
Realized that after changing CompStart to be in production, there was still a JSON file read error. This is because when the global boolean variable is_prod is False, the function get_prod_path returns a list that resolves to the path /devenv/feature_addons/startup_data_modifier_tool/config. This path no longer exists and will need to be updated.

However, when the is_prod variable is true, the list resolves to the path /config, which also doesn't exist. Rather, config is found at /devenv/config.

Changed is_prod back to false for now and updated the get_prod_path function to return a list that resolves to /devenv/config when is_prod is false.