Minor code change in generate_default_startup_data
In that function, the get_prod_path function in the helper module is called, which returns a list to resolve to the path of the JSON file in question. It used to be that the helper function didn't include the config folder, so the code in generate_default_startup_data would extend the list to include config and pass that list on.

However, somewhere along the way, the get_prod_path function was modifed to add config to the list, so extending the list to add config a second time was causing errors. This has been corrected and confirmed working.