Resolved the issue
Found the spot where the debugging info is printed out when returning from the add-a-startup-item menu. Since the edit-json menu is what leads to that other menu, the function json_editor is responsible for the call to the function add_startup_item in the cs_startup_add module. It was surrounded by a print statement which is now removed. The resolution was confirmed.