Creating empty commit for success of PyInstaller test
Tested out the PyInstaller tool within the folder /devenv/experimental_content/pyinstaller and ran the command "pyinstaller CompStart.py".

It created a build folder, a dist folder and a CompStart.spec file as expected. The project .gitconfig file is already configured to ignore those folders and that file.

Within dist is a folder called CompStart and that contains CompStart.exe as well as a folder called "_internal". This folder contains all the compiled data from PyInstaller.

Ran CompStart.exe and everything seemed to work fine. However, only the functionality to view the startup file was tested.

Will need to experiment with creating a single file rather than a one-folder bundle. According to the PyInstaller docs, the single file compiles everything to a single EXE but each run creates a runtime folder in /tmp (or Windows equivalent). This means that there is no persistent data. Will need to test if that can still work with the location of config since the config folder will be shipped separate to the PyInstaller bundle of compiled files.