Modified the Release Notes for v1.1-alpha
Decided to modify the file gh_info.md under /releases/v1/m1/CompStart-1.1-alpha/release-notes.

Although this is realistically just retroactive - meaning the actual v1.1-alpha package will not be recreated with this new gh_info.md file - the changes have a better style to them as well as a more helpful description. Additionally, the Release Tag section is now a link to the actual GH page for the release.