Added back the original Release Notes for v1.1-alpha
Took the contents of the gh_info.md file from commit 8a7130 and pasted it back to the gh_info.md file for the release 1.1-alpha.

Renamed the existing gh_info.md file with all the updates from commit 72c745, commit 8d2062, and commit 82ab3c to (new)gh_info.md.

This was done so the two can be compared side by side. As mentioned in commit 8a7130, the actual release package will not be updated. Therefore, the version of the file that a user sees will be the original content. While the updates to the file contents are beneficial, it may not make sense to retroactively update the file contents for a past release.