Corrected mistake in previous commit
In the previous commit, commit 170516, it was noted that the Release Notes - the file gh_info.md - for release 1.1-alpha was reverted back to its original content to match what's in the actual release package.

However, the content copied over was incorrect and was instead, the changed content found in the Release Notes for release 1.1-beta. The correct original contents were grabbed from commit 10fef7 and pasted in.