Moved the Python CLI tool to its own folder
As commented in issue #50, the /release/v1/m1/1.1-beta/CompStart folder should not contain any Python scripts but only the Python CLI tool as an executable. This executable is to be generated using the Python PyInstaller library. See issue #43 for more information.

Created a new folder called py-tool under /releases/v1/m1/1.1-beta and moved all the Python scripts over to that folder. Will use that folder to generate the executable.