Updated the instructions for v1.1-beta
Updated the instructions text file for release v1.1-beta to include the current full list of files and folders.

Confirmed that the instructions for the section How to install CompStart were still correct. Updated the section How to modify the startup data file to let the user know about the new Python CLI tool.

Finally, removed the instructions under that section that were included in release v1.1-alpha but kept the note about the default startup data and what it contains.