Moved the py-tool version of CompStart.exe
Realized that the PyInstaller-created folders build and dist have automatically been added to .gitignore, so within the py-tool folder for release 1.1-beta, the generated executable isn't committed to the repository.

While the CompStart directory for this release has a copy of the executable, so that anyone else working on this knows there the generated executable is, it was moved to the py-tool folder.