Added a new dependency file
Created a new Python file called cs_description.py in the dependencies folder. This file will hold a global string variable that will be the actual description or explanation of what CompStart is and how it works.

When option 1 is selected from the main menu, this variable will be used for the output.