Updated cs_helper.py
Added an import statement to import the new cs_description module and refactored the program_info function to use the CS_DESCRIPTION variable for what is printed on the screen.

Note that the import statement references the new module as cs_desc instead of cs_description. Even though, in the previous commit, the new module was named cs_description.py, to fit the naming convention of the other modules, the name will be changed.