Refactored the way the description is printed out
The current program description has 7 sections, so modified the CS_DESCRIPTION variable to add a "--" in between each section.

The program_info function was then refactored to split using that delimiter, and print out each section one at a time. Each section is printed as a page with a page number at the bottom. Finally, a prompt asking the user to press enter to continue is given. This prompt changes slightly in wording depending on if the last page or section is being shown, or not.