Removed the override for generating new JSON data
In the function generate_new_json_data in the cs_data_generate module, there was an if-block override. This override would check if the user chose to create a startup file with programs of their choosing.

If they chose this option, the override would tell them that isn't possible and a default startup file will be created. The variable used to determine which type of file to create was then modified so the default file was created.

Since currently, in the function generate_user_startup_data, there is already a statement printed that the functionality isn't available and a blank startup file is created instead, there was no need for the override.

Note: The blank startup file isn't a blank file. It's a startup JSON file with 0 startup items.