Started task 1 from issue #11 on GitHub
Added an if-block to the json_creator function to check if the user chose to create their own startup items. The if statement also checks to see if the user was ok with overwriting an existing file, if there was one. This way, if the user decided to change their mind and selected N to overwriting an existing startup file, they aren't asked to add startup items.