Added a check within case 2 in the main loop
In the main loop, case 2 is for creating a new startup file. After the function json_creator is called and returned, the boolean variable status_state is checked, and if it's False, it meant that there was an error and so a new JSON file couldn't be created.

In this case, a message to check the other error messages was printed. However, the function json_creator also returns False for that variable when the user chose to not overwrite an existing startup file.

Modified the if statement to check the returned status_message from that function. If the user chose not to overwrite, the status message will be "Skipped writing startup file", which now gets printed out instead of a message about checking error messages.