Refactored json_creator
Based on the comments in issue #11 on GitHub, refactored the code added to deal with adding startup items of the user's choosing.

Moved the code to before the function json_writer is called. Also, changed the code flow around so if the user enters an invalid number for the number of startup items to add, an error is printed out, as mentioned in the previous commit, and execution is returned to the main loop.