Started work on task 2 from issue #11 in GitHub
Created a for loop to loop through the number of startup items requested by the user, and for each item, call the add_startup_item function in the cs_startup_add module.

One thing to note is that the add startup item menu that the called function displays gives the option to save the startup item. This option makes sense when the user reaches this menu by selecting to edit the existing startup data, then add a new startup item to this existing data. The option doesn't make sense in the case where a user is adding startup items to create a new startup file.

Another thing to note is that since the code was refactored in the previous commit so the user is asked how many items they want to add before the call to the function json_writer, when the user creates a new startup item, the item's ItemNumber property is set based on the existing startup data file, rather than what should be a blank startup data file.