Started work on task 2 from issue #11 in GitHub
Created a for loop to loop through the number of startup items requested by the user, and for each item, call the add_startup_item function in the cs_startup_add module.

One thing to note is that the add startup item menu that the add_startup_item function displays gives the option to save the startup item. This option makes sense when the user reaches this menu via Main Menu > Edit the existing startup file > Add a new startup item. The option doesn't make sense when the user reaches this menu via Main Menu > Create a new startup file > Create a new startup file with programs of your choice.

Another thing to note is that since the code was refactored in the previous commit to ask the user how many items they want to add before the call to the function json_writer, new startup items are created with an incorrect `ItemNumber` value. When the user creates a new startup item, the ItemNumber property is set based on the existing startup data file, rather than what should be a blank startup data file.