Refactored the call to json_writer
In the function json_creator, moved the check to see if the file already exists to the beginning of the big if-block. Also, moved the print statement about creating a blank startup file to the inner if-block that starts with:

if not default_mode

Moved the call to json_writer to after that inner if-block. Finally, moved the second if-block - the one that's used to loop through and have the user create the startup items - to outside the inner if-block and after the call to json_writer.

This now allows for blank JSON data to be written to file first and then for the user to add each startup item. It also ensures the flow of printed statements is more clear and makes sense.