Tested all the error messages for the function json_editor
Ran tests to make sure all the possible error messages for the json_editor function get printed out correctly when the situation occurs. During the process, realized that after one error message, the loop didn't quit to return to the main menu.

Also, added a print statement for the json_pruner function to let the user know that the item they deleted was successfully deleted.