Started refactoring how adding a startup item is handled
As shared in a comment on GitHub for issue #11, the add_startup_item function displayed a menu that included the option to save the startup item. This worked when that function was called through editing the existing startup file, but not when called through creating a new startup file with programs the user chooses.

In considering how to handle this, it was decided to refactor the whole add a startup item functionality. Specifically, the option to add a new startup item was moved to the main menu and removed from the edit a startup item menu.

Additionally, a new function in the module cs_jsonfn was created called json_adder. This function will handle asking the user how many startup items they want to add, validate and return an error if that number is invalid, and loop through to call the add_startup_item function for each actual startup item.

The function isn't complete, but the code previously in the function json_creator to ask the user how many startup items to add and validate their answer was copied over.