Changed how json_adder works
Decided to make json_adder more similar to the function json_pruner. This way, it will simply do one task, which is to allow the user to add one or more startup items.

As a result, the function now takes a dictionary containing the full JSON data and returns a dictionary containing the full JSON data but updated to include the startup item(s).