Finished working on json_adder and refactored add_startup_item
Updated the add_startup_item function to take an optional integer. If it's supplied, the function uses this value to calculate the ItemNumber for the new startup item. If it's not supplied, it defaults to -1, and the function then uses the get_count_total_items function in the module cs_helper

Also finished working on json_adder for the following:

- it now only checks if the user entered a negative number and prints an error

- the loop for adding the startup items only occurs if the user entered a number greater than 0

- a call is made to get_count_total_items and that plus the for loop counter is passed to the call to add_startup_item

- if the user did enter 0, a statement is printed out that no startup items were added