Changed json_adder and add_startup_item to return a true copy of the JSON data
The function edit_startup_item uses the deepcopy function found in the copy module to return a true copy of the JSON data that it returns. Additionally, in the various functions that deal with updating or saving startup data, this method is used to create true copies.

The reasoning is that Python passes by reference for iterable objects, and while that can work in some cases, in other cases, it's better to create a true copy.

The functions add_startup_item and json_adder weren't using this method but now they are.