Finished work on json_creator
Updated the code for json_creator to now call json_adder when the user wants to add their own programs.

Added an if-block to check and skip the call if the user wasn't ok with overwriting the existing file.

Also added an if-block after the call to check if the returned startup data is different from the existing one. This will cover the scenario where the user changes their mind and enters 0 when prompted for how many startup items to add.