Adding new test JSON file
While testing out the functionality of the json_adder function, 2 new and valid startup items were added to the startup data file.

However, when the CompStart.bat script was run, there were some issues with starting all the programs. The new test JSON file contains all 5 programs so the script can be troubleshot at a later time.

Referencing issue #11 for visibility as will need to create a new issue on GitHub to work on this.