Minor change to the add a startup item menu wording
After the changes done in issue #11, the first option in the add a startup item menu didn't make complete sense. It was worded as "create a startup item", but this is now displayed on the heels of the user being asked how many startup items to add in the function json_adder.

Changed the wording to "create the startup item" since the user is expecting to create a startup item when they get to this menu.