Made the same wording change to another part of the add a startup item section
The function add_startup_item is coded to display an "error" message to users when they try to modify a startup item's properties before actually creating the startup item. The message was worded as "please create a startup item first". Changed this to "please create the startup item first". This follows the same change as the previous commit.