Added the ability to display a save reminder message
In the function user_menu_chooser, added a new optional boolean parameter to allow calling functions to specify if, when printing out the menu passed in, a message should be added regarding saving.

The message acts as a reminder to the user that there is no autosave feature and they need to save any changes before returning to the previous menu.