Refactored the add_startup_item function for the new save reminder
Updated the add_startup_item call to user_menu_chooser to pass in True for the save reminder parameter. Also refactored the call to use name parameters.

Finally, decided to pass in to allow for a quit option in the menu to make it more consistent with all the other menus. Kept the parameter in the actual user_menu_chooser declaration for potential future use.