Fixed an issue where the user skips adding a startup item
Added validation to the function json_adder when it calls add_startup_item to check if the user actually created a startup item or not.

Modified add_startup_item to check, when a user chooses to return to the previous menu, if they created a startup item. If not, a blank JSON object is returned.

The function json_adder checks to see if the returned object is blank or not. If it's not, then the rest of the function continues. If it is, then a message is printed stating the startup item was skipped.

Finally, added an item counter to the for loop in json_adder so that the counter could be used to pass to add_startup_item the correct value for the parameter last_startup_item.