Added validation to fix issue #46
In the function edit_startup_item_arguments_list, when the user chooses to delete an argument, they are presented with a list of the arguments and a cancel option.

However, after the user_menu_chooser function is called, the code was only checking to see if the user chose the cancel option. The user_menu_chooser function returns 0 if the user entered any invalid input. Refactored the code to check for this as well.