Refactored the set_start_dir function
As per the issue identified in issue #56, the set_start_dir function was only taking into account 2 of the possible 5 scenarios for determining how to set the starting directory to the correct CompStart folder.

All 5 scenarios are now taken into account and tested to be working via manual manipulation of the path_dirs_list variable within the function. This hasn't fully been tested in a production environment, but the manual testing showed the logic to be correct.

However, now that set_start_dir returns a boolean variable, all locations that call this function need to be refactored to account for this.