Created the initial PowerShell script
Called it PTGenerator.ps1 and added an intro comment describing what it does. Per the comment:

This script is meant to be used for releases and will assume the release folder has been created. As a result, only a "py-tool" folder will be created where everything will be copied and worked on.