Added validation for the release directory
Reworded the third question to use the terminology of release tag. Added an if-block to test that the full release directory path exists.

If it doesn't, added a loop to ask the user if they want to create the directory. Took the loop code from CompStart.ps1 so it will only exit when the user enters a valid response.

In this case, if the user chooses not to create the releases directory, the script exits. Otherwise, the folder is created.