Added code to copy over the necessary Python scripts
Added 3 Copy-Item lines to copy over the CompStart.py script, the dependencies folder, and all the Python scripts within that folder from /devenv to the newly-created release folder.

The script also sets the location back to the project root, which will be helpful for running this script in the terminal of VS Code.