Attempted to fix issue of folders not being deleted when clearing py-tools
Added 2 lines of code to the if-block for checking to see if there's anything already in the py-tools folder. Attempted to use command prompt to delete the folders.

The command prompt line "for /D %v in (*) do rd /s /q %v" works when run in the terminal after changing directories to the py-tools folder.

However, using the PowerShell method of Start-Process to call command prompt and pass in the line above doesn't work. For now, the code has been commented out.