Copied over the initial code
Copied over the initial code form PTGenerator.ps1 that deals with creating the release folder. Cleaned up the code so it only has the lines for creating a new release folder.

However, currently only the full release folder path is tested for and created. This needs to be modified to also test for the major version and minor version folders.