Added checks for the release major and minor folders
Created new variables and renamed one variable to allow for the breakup of the path between the major version folder, the minor version folder, and the actual release folder.

Added the same if-block checks to see if the major and minor version folders exist.

Tested it out with multiple scenarios and the script correctly outputs when one of the folders doesn't exist.