Set up testing of each of the path components
Added 3 boolean variables and a tuple of 3 booleans to use. Added 3 if-blocks to test for the major version, the minor version, and the release folders. The boolean variables are updated based on the tests and the tuple is created from those booleans. This will quickly allow future code to determine at which point of the path, if any, do new folders need to be created.