Tested that everything works
Ran CopyReleaseContent.ps1 to test that the code works. Had to make a few adjustments including testing to see if the CompStart and release-notes folders already existed.

Noted while testing that if the content in CompStart already exists, while the Copy-Item command will overwrite an existing file (such as CompStart.ps1), it will throw an error with a folder that already exists.