Created a new PowerShell cmdlet to set the proper starting directory to the project root
The function was created using GitHub Copilot. It is the same function and logic as the function set_start_dir in the Python module cs_helper.

The function code was pasted into GitHub CoPilot and asked to convert it to PowerShell code. The resulting code was then copied to this new script.