Modified the Python function set_start_dir in the cs_helper module
The function that Copilot returned in the last commit used camelCase for the variables.

After a discussion on Reddit (see https://www.reddit.com/r/PowerShell/comments/1hj0iyh/camelcase_vs_pascalcase/), the plan was to change all the PowerShell scripts to use camelCase within functions.

However, it was decided to stick to the convention followed so far which is to use PascalCase for all variable, parameter, and function names.

The set_start_dir function was quickly modified to take a parameter and then the function was converted to PowerShell again using Copilot. While doing this, it made more sense to keep things that way - to have set_start_dir use a passed in directory name, rather than having it hard coded.

The main CompStart.py script was then modified to use a global variable for the starting directory and pass that in to the call to set_start_dir.