Minor correction to how the path is split
The code [System.IO.Path]::DirectorySeparatorChar, which was used to split the path to the passed in directory, was giving an error. This is because the script needed an escape character since, in Windows, the path separator is "\" which is the usual escape character.

This was modified to add a "\" in front of [System.IO.Path]::DirectorySeparatorChar and use that as the split character.

This was tested to work.