Modified when the variable $ProjectRootPath is initialized (see details)
Note: This commit was reverted because one of the changes related to this commit was not included. It has now been included so all the changes described below are being committed again.

While testing, realized that for all 3 release automation scripts, the $ProjectRootPath variable needs to be initialized after the start directory is correctly set.

Also made a minor name change in new-release.bat as it was set to call "CopyReleaseContents.ps1", but the third script is named "CopyReleaseContent.pst1".

Tested the changes by running the Batch file new-release.bat and it all works.