Minor changes to the GeneratePythonTools.ps1 and CreateReleaseFolder.ps1 scripts
Added the same ending as CopyReleaseContents.ps1 where the cwd is changed back to the project root and the user is alerted to that.

Also added a sleep timer variable to GeneratePythonTools.ps1 that's the same as in the other two scripts, and Updated one alert to the user.

Finally, changed the sleep timer for CreateReleaseFolder to 2 seconds to match the other two scripts.