Added to the GenerateReleasePackage.ps1 script
Partially with the help of GitHub AI and partially by copying from the release automation scripts, wrote up the code for the following tasks:

- get the release major version, minor version, and release tag from the user
- check for the existence of the release folder
- check for the existence of the CompStart folder, release-notes folder, and instructions.txt file in the release folder
- check for the existence of the packages folder

In all 3 cases, the user is alerted if the checks fail. However, for the last check, this needs to change to have the script create the package folder.