Added release directories for v2 to .gitignore
While testing out the GenerateReleasePackage.ps1, realized that a release folder will need to exist with all the content. Created release 2.1 using the release automation scripts.

Added the /releases/release-versions/v2 folder to .gitignore so those directories and files won't be synced but can still be used in testing the package automation script.