Updated folder names for consistency
Realized that some folders used a hyphen between words (such as /releases/release-versions) and others used an underscore (such as /devenv/experimental_content).

The underscore was initially used because Python prefers snake_case and that was used for the Python script names. However, later, the hyphen was used for folder names as it looks cleaner.

This just makes all the folders under CompStart use a hyphen when needed for consistency.

The exception to this is the most recent feature branch features/create_release_package. However, going forward, feature branch names and folders will use the hyphen.