Updated CompStart scripts under the powershell-installer folder
The powershell-installer folder from a previous feature branch was moved to /devenv/features/cs-installer in a previous commit.

Within that folder is a subdirectory called installer-files, which contained the CompStart scripts. They have been updated to the latest version, which includes CompStart.py.

Also, the script startup.ps1 was removed from installer-files as it was a duplicate of CompStart.py.