Added a function to create the CompStart folder
Added a function New-CSFolder that will create the CompStart folder. It takes an optional string parameter that would function as the parent path for the folder. If it's not supplied, or the supplied parameter isn't a valid path, the folder is created at %LocalAppData%.

Also created 2 global variables. Made the directory separator variable a global one and add a global sleep timer variable.