Made the folder path variable a global one
Started writing a function to copy the CompStart files and realized that the path to the CompStart folder would either need to be passed to every function or made a global variable.

Decided to make it a global variable. When New-CSFolder is called, if a valid path is provided, the global variable is updated. This will ensure other functions will always have the right path to use.