Added Batch comments to CompStart.bat
Used VS Code Copilot to generate Batch comments in the style of PowerShell comment help complete with a .SYNOPSIS, .DESCRIPTION, .PARAMETER, .EXAMPLE, and .NOTES section.

Took the .Notes section from CompStart.ps1. The generated comments were all in the style of REM comments, but modified them so the lines with the categories (.SYNOPSIS, etc.) were in the style of the double colon comments and the actual content was a REM comment.

Copilot also generated REM comments for each line of code.