Finished creation of the prodenv directory
Moved over both release version folders and package artifacts for CompStart-1.1-alpha and CompStart-1.1-beta to the /prodenv folder.

Created a /prodenv/packages and a /prodenv/releases directory. Moved what was in /releases/versions to /prodenv/releases as well as what was in /packages/versions to /prodenv/packages.

Additionally, moved both package and release READMEs to the respective prodenv folders.

Finally, removed the /releases and /packages folders.