Deleted the version-specific release note MD files
Realized that the release version-specific release-notes.md and instructions.txt files are already stored in the release folder. Therefore, it was unnecessary to store a copy in /prodenv/assets.

Additionally, renamed the /prodenv/assets/release-notes folder to /prodenv/assets/release-assets.

Finally, moved the "template" instructions.txt and release-notes.md files to this new release-assets folder.