Added a new function Start-Release
Created a new function called Start-Release which was largely suggested by GitHub Copilot. The comment help and the parameter block have been added.

However, the actual function hasn't been coded. The function will call Add-MajorVersion, Add-MinorVersion, and Add-ReleaseVersion. The function will also do this for both the /prodenv/releases path and the /prodenv/packages path.

Need to decide if this function will check for the existence of the release folders first, before calling the other functions, or another function will be used to do this.

However, the overall idea with the DeployRelease script will be to make everything as functional as possible.