Created three new functions
Created three new functions called Set-MajorVersionPath, Set-MinorVersionPath, and Set-ReleasePath. These functions will do the actual work of checking for existing folders and creating them if necessary. The checks will be done for both /prodenv/releases and /prodenv/packages.

For example, for release "1.1-alpha", the first function will check if there's a "v1" folder for both the releases and packages directories. The second function will check if there's an "m1" folder. The final function will check if there's a "1.1-alpha" folder but only for the releases directory.

The comment help for all three functions was also added along with the parameter blocks.

Finally, the functions Add-MajorVersion, Add-MinorVersion, and Add-ReleaseVersion were renamed to have Folder at the end of the existing name. This was done to create clarity as those functions only create the relevant folder.