Started working on the main portion of the DeployRelease script
In the Out-of-Code Insights annotations (see last 2 commits), the following reply was added as a next step:

----------------------------
All 4 scripts have code at the beginning that calls the Set-StartDirectory (renamed to Set-ProjectRoot) function, checks the return result, and quits if necessary. While the code lines for this was only copied from the CreateReleaseFolder script, the code could probably be set as the first thing DeployRelease does. Or, the code could also be set to its own function but that function is called first.
----------------------------

After all the functions, there's a comment for the start of the main script. That portion of the copied code was then added as the first thing the script does. Afterward, the script then calls the function Start-Release.