Changed a parameter in the function Start-Release
While testing the code, realized that the third parameter for the function Start-Release was called ReleaseVersion and the idea was that the caller would pass in the full release version.

However, the code in the old automation scripts get the major version, the minor version, and the release tag from the user. The full release version is then created from the responses.

The third parameter in Start-Release was changed to be the release tag, which will allow the function itself to determine the full release version.