Refactored the other two parameters of Start-Release and added code to determine the full release version
Renamed the other two parameters of Start-Release from $MajorVersion to $ReleaseMajorVersion and from $MinorVersion to $ReleaseMinorVersion.

Added a line to the .Notes section of the comment help for the function to show an updated date, which is when the function comment help was updated.

Removed the Mandatory parameter attribute from the $ReleaseTag parameter.

Finally, copied the code from the commented-out code that creates the variable $ReleaseFullVersion and populates it with the full release version, such as "1.1" or "1.1-alpha".