Committing previously ignored folders and files
After the change to .gitignore, this commit includes the files within .vscode and the files within the py-tools folder for release 1.1-beta (that is, /prodenv/releases/v1/m1/1.1-beta/py-tools).

Specifically, it includes launch.json and settings.json from the /.vscode folder. It also includes the CompStart.spec file and the subdirectories build and dist, all found within the py-tools folder.