Undid the minor modifications to the ReleaseDetails dictionary
Minor changes were made 4 commits back to the ReleaseDetails script variable dictionary. Rather than defining the properties of the dictionary on variable initialization, the function Get-ReleaseDetails was changed to use the Add method of HashTables to add each property.

This was changed back so all four dictionary properties - MajorVersion, MinorVersion, Tag, and FullVersion - are defined on variable initialization and set to a blank string.

This will make it easier to see, at a glance, the properties of the ReleaseDetails dictionary without needing to read through the code.

The same approach will be taken for all the other script variable dictionaries.