Added and updated the properties for the script dictionaries
For the PathVars dictionary, added properties for the various paths that will be needed to be tracked. Updated some of them based on the current directory structure, but more changes may need to be made.

Organized the properties for all 4 dictionaries so there are blank lines to group together similar properties. For example, in the FileNames dictionary, all 4 properties related to the CompStart files - .bat, .ps1, .py, .exe - are grouped together.

Made some other minor modifications.