Adjustments to the dictionary properties for PathVars and FolderNames
Added a new FolderNames property called ParentInstallLocation and set it to LocalApplicationData. Renamed the PathVars property InstallationLocationParent to ParentInstallPath and used the new ParentInstallLocation property to set it. This will allow for easier updating if the default installation location will change.

Also added a property to PathVars called ProjectRootPath and set it to a blank string for now.