Created a new function called Update-PathVars
This function goes through and updates all the PathVars dictionary properties. Set the main script to call this function after Get-ReleaseDetails and also to use Get-Location to update the PathVars property ProjectRootPath.

In the new function, started assigned and updating the first 19 or so properties of PathVars, which covers the following directories:

/devenv
/devenv/config
/devenv/dependencies

/prodenv
/prodenv/assets
/prodenv/packages
/prodenv/releases

/prodenv/assets/cs-installer
/prodenv/assets/release-assets

/prodenv/packages/v<#>
/prodenv/packages/v<#>/m<#>

/prodenv/releases/v<#>
/prodenv/releases/v<#>/m<#>
/prodenv/releases/v<#>/m<#>/<full-release-version>

Tested using "Write-Host ($Script:PathVars | Format-Table | Out-String)" and confirmed all the paths are set up correctly.