Added as comments the rest of the path variables
Copied over from the commented-out code the rest of the path variables as well as the relevant properties from the PathVars dictionary. Surrounded all of them within a comment help style section "<#..#>" and added the Write-Host line to the end of the function.

Left the commented-out section at the end of the main script that stays "Temporary holding place for copy-pasting of all the script variables needed for the script" for now.