Refactored the functions Set-MajorVersionPath and Add-MajorVersionFolder
Renamed the first function to Set-MajorVersionPaths. Refactored the comment help for both functions to the fact that the functions now have the script dictionaries to use, so they don't need the same parameters as before.

The Set-MajorVersionPaths function checks for a major version folder at both /prodenv/releases and /prodenv/packages. The Add-MajorVersionFolder function now takes one parameter, a boolean that specifies whether to create the folder under releases or packages.

The output to the user was also updated to be more concise but still explanatory.

Finally, the function placements were reorganized.